Complete rework of the hashing thread class, which now (among other improvements) handles problems more gracefully and supports multiple threads should you want to use that.
One change is the fact that PartFiles are moved to the front of the queue rather than the back, so that they get hashed sooner.

Most of the logic was moved from CKnownFile::CreateFromFile, which made it possible to improve upon the original design.  However, instead of one, 3 mutexs are now used, because of the added complexity though care has been taken to insure as short locking times as possible.

The class has also been throughoutly documented.


git-svn-id: http://amule.googlecode.com/svn/trunk@1547 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff