Clean-up of systray. I've run it through astyle so that at least it uses normal indentation and also done other clean-ups and simplified some code.
This should also fix the problems with the systray fucking up rate-limits (now calls ForceUlDlRateCorrelation() rather than doing it's own checking). Also did a slight rework of the actual menu.

These patches enabled the usage of the PublicIP requests, which the systray now makes use of to get the client IP, rather than checking the net-interfaces.


git-svn-id: http://amule.googlecode.com/svn/trunk@1561 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff