New files, 	NetworkFunctions.h 	NetowrkFunctions.cpp
Handling the network string/bytes conversion of ips and hostnames and the generic hostname solving thread, using no more native code but a funny way to wrap using wx native functions (be sure to check it out for real fun!).

Also cleaned a lot of code. And probably fixed some of the ipfiltering bugs. Probably not all.

Improved network a lot. Check it out :)


git-svn-id: http://amule.googlecode.com/svn/trunk@2343 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff