lfroen, I know being high on drugs is ok - I did it lot of times myself. Just try to not to code on that state, ok?
a) WTF a BarShader on a core? It's a GUI file and stuff!
b) WTF sending a line to GUI's? Every GUI must build it's own bar with the style it feels like, based on file parts!
c) WTF using non-wxBase things on a wxBase app? Because, you know, people will try to comiple on wxBase and won't be able!
d) WTF breaking CVS just before a release? Because, you know, it doesn't compile! wxBase has no wxRect and such!


git-svn-id: http://amule.googlecode.com/svn/trunk@2514 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff