This commit have double function:
a) Fixed the crash on non-std locales because of date/time functions. This was happening because of a bug on
wxWidgets, so I report it and also make we use ISO data/time so we don't crash anymore.

b) Get rid of more ugly unicode2char stuff by making the logfile a wxFFileOutputStream. Also speedup on writing to
logfile.


git-svn-id: http://amule.googlecode.com/svn/trunk@3004 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff