Fixed the handling of m_toresolve list of hostnames to resolve.  In AddToResolve, push it on the front because when the CAsyncDNS thread completes and calls OnHostnameResolved, it assumes the just-resolved hostname is the one on the front.  Similarly, AddToResolve's error handling acts like the just-pushed hostname is on the front.
Later, in OnHostnameResolved, break out of the loop which spawns the next
thread to resolve the next hostname when it successfully runs the thread.
Otherwise, depending on the order of thread scheduling, we could get many
duplicate threads running and resolving the same hostnames, possibly
resulting in an assert firing.

Kry and everyone, check my logic on this.  I may have misunderstood the
code.


git-svn-id: http://amule.googlecode.com/svn/trunk@3314 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff