Fixed a small bug in UTF82unicode(), reported by Mr Faber.
The patch in StringFunctions.h avoids problems when UTF82Unicode() might
return NULL, e.g. CFile.cpp:698, where it has to process a filename that
is in UTF-8, and must not fail upon return. The problem manifest in non-
unicode builds.

In non-unicode builds, wxConvUTF8.cMB2WC() will return NULL, if it
cannot convert an UTF-8 string to char *, what is likely to happen. In
unicode builds, I have also replaced wxConvUTF8.cMB2WX for
wxConvUTF8.cMB2WC. Although the behavior in unicode builds would not
fail, i did that for consistency with the non-unicode usage.


git-svn-id: http://amule.googlecode.com/svn/trunk@3413 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff