Got rid of sscanf in NetworkFunctions.h. This has caused a crash once, reported by Mr Faber.
The patch in NetworkFunctions.h, is to address an issue relating
amuled. It seems to be chocking on the sscanf. Worst, besides chocking,
it locks with high cpu usage, but this is related to not beeing able to
execute the FatalErrorHandler() and must be addressed later. This way we
avoid the sscanf and hopefully avoid a crash.


git-svn-id: http://amule.googlecode.com/svn/trunk@3414 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff