Improved error-handing in the ReinitializeNetwork function. If any of the ports have an invalid value (from 1025 to 2^16 - 1), then we select a random value so that we at least can run, even if with reduced efficiency due to firewalling.
The old code would simply return which would cause the sockets to be left uninitialized and cause crashes during startup.

Please try to avoid such cases where we cant start, unless it is absolutely necessary (for instance wx2.4.2+gtk2) as it will leave most users unable to fix it. If possible, select a safe value as here.


git-svn-id: http://amule.googlecode.com/svn/trunk@4112 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff