We just have to not get used to one RSB per year ;)
We should not add a client to the deletion queue directly, NEVER. It has to be done via Safe_Delete(). I added an
assert there now so this never happens again.

For those familiarized with RSB: client deletion was being done directly, without Safe_Delete, so socket events were
happening right on deletion.


git-svn-id: http://amule.googlecode.com/svn/trunk@4690 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff