Adding checks to all places in aMule (minus kad-code) that uses tags without checking if the type matches the expected type. This could previously result in crashes or wrong values when dealing with corrupted or maliciously crafted tags.
Also changed the asserts in the CTag::Get* 
functions to checks, so that we don't risk
crashing in case we miss a check. 

Lastly, the action taken on unknown tag-types
has been changed from simply printing a warning
to throwing an exception, due to the fact that
we are unable to safely proceed from the tag,
since we do not know the length of the tag and
therefore cannot skip past the data contents.


git-svn-id: http://amule.googlecode.com/svn/trunk@5073 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff