Renamed the ReadHash16 and WriteHash16 to simply ReadHash and WriteHash. ReadHash was changed to return a CMD4Hash object and WriteHash was changed to take a const reference to a CMD4Hash object.
Also removed the implicit casts to unsigned char* from 
CMD4Hash and added operator[] instead. The array can
still be accessed through GetHash().

Also changed the client-credits struct to use CMD4Hash
objects rather than char arrays and removed the "pack"
pragmas, since we do not read or write directly to/from
the struct any more.

CTag was also changed so that it uses CMD4Hash objects,
and a few rarely used functions in CPacket were moved to
the cpp file in order to reduce the need to include
files in the header.


git-svn-id: http://amule.googlecode.com/svn/trunk@5158 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff