Un-translating some strings in muuli.
Unfortunately wxDesigner maintains the same flag for translating the
value of a text contron and for its tooltip text, so it isn't possible
to mark the text "not-to-be-translated" and still have its tooltip 
translated (this would be needed e.g. for the serverlist update URL,
where the URL doesn't need translation, but its tooltip does).


git-svn-id: http://amule.googlecode.com/svn/trunk@5319 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff