Fix for filenames that has a length between 60 and 70, short name == full name + " ..."
Though php_core_lib.cpp is not in POTFILES.in, using
wxT() instead of _() is encouraged, since it is faster
(no translation lookup required), and will save some
troubles when the string inside the apostrophes matches
a string that has a translation.
For example, it was "..." instead of " ...", short name
would render on hungarian a <truncated filename> + "Tallóz".
Pretty much not what we wanted :)


git-svn-id: http://amule.googlecode.com/svn/trunk@5358 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff