Proper config location on Mac and Win.
In case an old configuration is detected, it will be moved silently to the new location,
unless the new location contains a valid configuration. When moving configuration,
config file will be edited to make sure every path in it is still valid after the move.

NOTE: config relocation is only done by amule(d)


git-svn-id: http://amule.googlecode.com/svn/trunk@5416 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff