Adding new CMuleUDPSocket class.
This class takes care of the common tasks of sending
and receiving (including error handling) and
interfacing with the upload throttler. It also
includes a fallover recovery for when the socket
die (error 4).

The class does currently not support rebinding of
sockets, but it is trivial to add.

This patch affects both the ClientUDP and ServerUDP
sockets, so both are now controlled by the UBT.

This commit should resolve bug #489:
 http://bugs.amule.org/view.php?id=489


git-svn-id: http://amule.googlecode.com/svn/trunk@5603 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff