For the Mac, instead of gunzipping via wxZlibInputStream, use the zlib functions (gzopen, gzread, gzclose) directly.  This is to allow for gunzipping on Mac OS X 10.3.x (Panther), which only has zlib 1.1.3.  With that version of zlib, wxZlibInputStream doesn't gunzip.  Since our default server.met is gzipped, that's a significant problem.
The new code should work just as well on non-Mac platforms.  For the time being,
since we're close to release, I decided not to disturb the behavior on non-Mac
platforms.  In the future, others should test the new code and we should use it
for all platforms if it works well everywhere.


git-svn-id: http://amule.googlecode.com/svn/trunk@6219 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff