Partial cleanup of the HTTP download thread. The dialog should now also work on Mac, since it now uses events to signal changes in states.
Use of a dialog window is now handled via a 
constructor parameter, instead of having multiple
thread classes, and the dialog class has been 
moved to the CPP file.

NOTE: This also includes a *VERY CRITICAL* update
of the download dialog, and it is recommended that
all who use it upgrade immediatly.


git-svn-id: http://amule.googlecode.com/svn/trunk@6390 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff