Automatic removal of unneeded headers. Quite buggy and might not compile on several systems.
If it doesn't compile on your system, and you know the header to add, 

PLEASE INCLUDE  // Do_not_auto_remove

after the include irself, so it will not be autoremoved next time. 

If you don't know which on to include, post on testers forum and let me add it (or someone).

Let's break.



git-svn-id: http://amule.googlecode.com/svn/trunk@6844 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff