Changed data types to clear warnings at /W4 (mex level) on MSVC. Also cleared the MD5_ functions to use standard memset/memcpy. While at it, removed the type defines and replaces the proper instances.
Clearing warnings is fun, because you find bugs like the one on TextFile.cpp CTextFile::WriteLine. Notice how wxEmptyString has a implicit conversion to wxChar* that will always be bool "true", meaning we return that we wrote the line correctly when it failed because the function returns bool, not wxString.

git-svn-id: http://amule.googlecode.com/svn/trunk@8338 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff