Added preferences option 'Allocate full size'
When enabled, new partfiles will try to allocate full size, so the download later won't be stopped
because of lack of disk space. If an error happens while allocating, the download will be stopped
and an appropriate message is logged.

We don't set the partfile status to:
- PS_ERROR, because that can't be cleared, so the download cannot be started
- PS_INSUFFICIENT when allocation fails because of lack of disk space, because on the next processing
of the download queue this state would be cleared and the download started, if free space is enough to
start it, but less than full file size.

Failure in allocation in MSW will result in a 0-size file, which is not set to sparse, so the download
may fail if the downloaded chunk is far from the beginning of the file.

Current preallocation method on MSW works only if the file system does not automatically set all files
to sparse, i.e. it doesn't work with amule-msw under wine on linux using ext3.

We also preserve the value of 'Start new files paused', because user might change it between the creation
of a new download and the start of the download - preallocation might take some time to complete.

git-svn-id: http://amule.googlecode.com/svn/trunk@8542 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff