Fix for improper handling of EC connection failures.
First of all, thanks to wuischke for his research on the problem, which made it
easy to create a real fix.

The problem was that callers and called function disagreed on the meaning of the return value.
I simplified the code, but I did not test it under wxWidgets earlier than 2.8.8, so I just left
the old way in case it would not work.

Some documentation update is also needed...

git-svn-id: http://amule.googlecode.com/svn/trunk@9004 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff