Simplified --enable-* option handling.
Added functions to declare and query --enable-* flags, remembering their default
value throughout the configure run. This eases changing the default behaviour
(only one place has to be modified) and prevents wrongly parsing/displaying their
values because of a forgotten update. For example the default value for --enable-xas
was `no', but for the final summary the fallback value was `yes'. This didn't
cause any problem though, as $XAS was always set.

git-svn-id: http://amule.googlecode.com/svn/trunk@9925 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff