Fixed crash in CClientList::ProcessDirectCallbackList that had been around for ages
aka CUpDownClient::ClearDownloadBlockRequests, see http://forum.amule.org/index.php?topic=17604.0
Problem was that a client already scheduled for deletion was added to the DirectCallbackList and later accessed after it was deleted. It only happens with some uploading clients, probably with a special timing.
Final confirmation is missing though because the problem went away meanwhile for Supersnail. But I'm confident that was it.

git-svn-id: http://amule.googlecode.com/svn/trunk@9989 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff