Disable EC notification completely for clients not supporting it.
Just got a debug assertion from it (running amuleweb) after deleting and then readding the same file.
After deletion file doesn't get removed from m_dirty_status because packet is not sent, and then it asserts when it gets readded. 
I wonder why it was active for non-notification clients in the first place.

git-svn-id: http://amule.googlecode.com/svn/trunk@10026 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff