Make CECTag::TagList a list instead of a vector
Pushing back to a vector leads to resizing and copying over which is ugly for complex tags.
This required replacing a few (well - a lot) of loop iterators.

Drive-By-Fix: "pause/cancel/resume all" in aMulecmd was broken for a while (since switching Partfile tags to ECID), but in 2.2 it applied to one file only instead of all because of a break placed in the wrong brace level.

git-svn-id: http://amule.googlecode.com/svn/trunk@10085 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff