Simplify queue score logic
1) removed queue "rating" (used only in client detail dialog)
2) removed option to not upload full chunks
This was a CPU hog, since it required for each packet prepared to compare queue score of the client with the scores of all clients in the queue (which all had to be calculated first).
Now there is no more queue score required for a downloading client.

The goal is to get an ordered upload queue with each client knowing its position without any calculation at all. Queue order will be updated when necessary only - when an upload slot has to be assigned. This will fix CPU load problems with sort-by-queue-pos in client list control an with EC.

git-svn-id: http://amule.googlecode.com/svn/trunk@10259 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff