Try to find the closest node alive.
In a find lookup, if the best two nodes are dead/unreachable, then try to
discover more close nodes before using our other results. The reason for
this is that we may not have found the closest node alive, due to results
being limited to two contacts, which could very well have been the
duplicates of our dead closest nodes.

Note that this is just a quick intermediate fix to make certain Kad lookups
more reliable, improving the results in some cases [ based on research from
http://www-users.cs.umn.edu/~hopper/kad.pdf ]

git-svn-id: http://amule.googlecode.com/svn/trunk@10273 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff