Get rid of the strict-aliasing warnings with GCC 4 and optimizations enabled
The check is only for the GNU C++ compiler, because that will use builtin memcpy()
with -O2 (which also makes strict aliasing enabled), and thus it won't cause any
runtime penalty.  To actually get a call to memcpy() you need to disable optimization
and manually enable strict aliasing ;)

Also the check for the GNU C++ library has been disabled because its result is not used.

git-svn-id: http://amule.googlecode.com/svn/trunk@10703 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff